Renamed SearchBarIOS component to SearchBar
